refactor(launcher): split reducer.rs into mod.rs + tests.rs (task #182 PR-A)
Pure mechanical extraction. Zero logic changes. The launcher reducer
file was 3877 lines (71% of which was tests). Split:

- agentmux-launcher/src/reducer.rs → agentmux-launcher/src/reducer/mod.rs (1114 lines, prod only)
- New: agentmux-launcher/src/reducer/tests.rs (2763 lines, all tests)

mod.rs ends with `#[cfg(test)] mod tests;`. tests.rs starts with
`use super::*;` (de-indented from the prior `mod tests { ... }`
inline block).

This is the first bite of task #182 (modularize). It DOES NOT split
the tests further by domain — that's a follow-up. The point of this
PR is to halve the prod-side file's effective size for navigation
and future per-domain handler splits (PR-B in the task description).

154 launcher tests pass. No new tests, no removed tests.
